requirements.txt
torch
torchvision
numpy
scikit-learn
tqdm
django
pillow
